api: modify settings api

+++ mandarin-common/src/main/java/org/a3/mandarin/common/util/ValidateUtil.java
@@ -21,7 +21,7 @@ public class ValidateUtil {
         return true;
     }
     
-    public static Boolean validateNumber(Integer number){
+    public static Boolean validateNumber(Double number){
         // TODO
         return true;
     }
